Generate labels/annotations for Tekton results earlier
We have to set labels/annotations sooner than in exit plugins which run
as finally task, because tasks in finally can't update tekton results.

* CLOUDBLD-8890
