+++ atomic_reactor/plugins/pre_resolve_remote_source.py
@@ -25,6 +25,7 @@ from atomic_reactor.constants import (
     REMOTE_SOURCE_TARBALL_FILENAME,
 )
 from atomic_reactor.dirs import BuildDir
+from atomic_reactor.metadata import annotation_map
 from atomic_reactor.plugin import PreBuildPlugin
 from atomic_reactor.util import is_scratch_build, map_to_user_params
 from atomic_reactor.utils.cachito import CFG_TYPE_B64
@@ -63,6 +64,16 @@ class RemoteSource:
             return REMOTE_SOURCE_JSON_FILENAME
 
 
+def transform_results(results):
+    remote_sources_annotations = [
+        {"name": remote_source["name"], "url": remote_source["url"]}
+        for remote_source in results
+    ]
+
+    return remote_sources_annotations
+
+
+@annotation_map('remote_sources', transform_results)
 class ResolveRemoteSourcePlugin(PreBuildPlugin):
     """Initiate a new Cachito request for sources
